pyproject.toml
[project]
name = "local-web-services"
version = "0.6.0"
description = "Run AWS CDK applications locally - accelerate development with agentic code editors"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["aws", "cdk", "local", "development", "lambda", "dynamodb", "sqs", "sns", "s3", "ssm", "secretsmanager", "agentic", "ai", "code-editor", "vibe-coding"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "typer>=0.9.0",
    "watchdog>=4.0.0",
    "aiosqlite>=0.20.0",
    "croniter>=2.0.0",
    "rich>=13.0.0",
    "python-multipart>=0.0.9",
    "pyjwt[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "docker>=7.0.0",
]

[project.urls]
Homepage = "https://local-web-services.github.io/www"
Repository = "https://github.com/local-web-services/local-web-services"
Issues = "https://github.com/local-web-services/local-web-services/issues"

[project.scripts]
ldk = "lws.cli.ldk:app"
lws = "lws.cli.lws:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lws"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "radon>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "radon>=6.0.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "C90"]

[tool.ruff.lint.isort]
known-first-party = ["lws"]

[tool.ruff.lint.per-file-ignores]
"src/lws/api/gui.py" = ["E501"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pylint.main]
load-plugins = []

[tool.pylint."messages control"]
disable = [
    "E0401",   # import-error — false positive under uvx (no venv)
    "W0718",   # broad-exception-caught — suppressed globally per user decision
    "R0801",   # duplicate-code — handled by `make cpd`
    "C0301",   # line-too-long — handled by ruff/black
    "C0302",   # too-many-lines — handled by ruff
    "C0114",   # missing-module-docstring — not enforced in this project
    "C0115",   # missing-class-docstring — not enforced in this project
    "R0903",   # too-few-public-methods — dataclass/protocol pattern
    "R0902",   # too-many-instance-attributes — config objects
    "R0913",   # too-many-arguments — structural
    "R0917",   # too-many-positional-arguments — structural
    "R0904",   # too-many-public-methods — provider pattern
    "R0911",   # too-many-return-statements — structural
    "R0914",   # too-many-locals — structural
    "C0103",   # invalid-name — naming conventions differ from pylint defaults
    "W0603",   # global-statement — used sparingly, intentional
    "W0621",   # redefined-outer-name — rare, intentional
]
